isbn수정

+++ master/oneSentence_master/src/main/java/com/one/sentence/onesentence/dao/IOnesentenceDao.java
@@ -14,23 +14,23 @@ public interface IOnesentenceDao {
 	public List<ShowOnesentence> selectOnesentenceList();
 	public List<ShowOnesentence> selectOnesentenceListByuserIdx(int userIdx);
 	public List<ShowOnesentence> selectOnesentenceListForLiketo(int userIdx);
-	public ShowOnesentence selectOnesentenceByoneSentenceIdxAndIsbnMy(int userIdx, int isbn);
-	public ShowOnesentence selectOnesentenceByoneSentenceIdxAndIsbnFollowing(int userIdx, int isbn);
-	public List<ShowOnesentence> selectOnesentenceListByoneSentenceIdxAndIsbnOther(int userIdx1,int userIdx2, int isbn);
+	public ShowOnesentence selectOnesentenceByoneSentenceIdxAndIsbnMy(int userIdx, long isbn);
+	public ShowOnesentence selectOnesentenceByoneSentenceIdxAndIsbnFollowing(int userIdx, long isbn);
+	public List<ShowOnesentence> selectOnesentenceListByoneSentenceIdxAndIsbnOther(int userIdx1,int userIdx2, long isbn);
 	
 	public ShowOnesentence selectOnesentenceByoneSentenceIdx(int oneSentenceIdx);
 	public OneSentence selectOnesentenceModel(int oneSentenceIdx);
-	public int updateOneSentence(int oneSentenceIdx,String oneSentence, String page, int isbn);
+	public int updateOneSentence(int oneSentenceIdx,String oneSentence, String page, long isbn);
 	public int deleteOnesentence(int oneSentenceIdx);
 	
 	public int insertBook(Book book);
-	public Book selectBookByisbn(int isbn);
+	public Book selectBookByisbn(long isbn);
 	
 	public int upUserPoint(int userIdx);
 	public int downUserPoint(int userIdx);
 	
 	public int insertHashtag(Hashtag tag);
-	public int findOneSentenceIdx(int userIdx, int isbn, String oneSentence);
+	public int findOneSentenceIdx(int userIdx, long isbn, String oneSentence);
 	public List<String> selectHashtagList(int oneSentenceIdx);
 	
 	public int selectLikeTotal(int oneSentenceIdx);
